refactor: PaymentConfirmPage가 recoil에 의존하지 않도록 구현
